refactor: Time::Iterator -> TimeIterator
This avoids possible namespace collisions with Time.
