+++ README.md
@@ -1,8 +1,8 @@
-# Time::Iterator
+# TimeIterator
 
 TODO: Delete this and the text below, and describe your gem
 
-Welcome to your new gem! In this directory, you'll find the files you need to be able to package up your Ruby library into a gem. Put your Ruby code in the file `lib/time/iterator`. To experiment with that code, run `bin/console` for an interactive prompt.
+Welcome to your new gem! In this directory, you'll find the files you need to be able to package up your Ruby library into a gem. Put your Ruby code in the file `lib/time_iterator`. To experiment with that code, run `bin/console` for an interactive prompt.
 
 ## Installation
