--- lib/time/iterator/version.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class Time
-  module Iterator
-    VERSION = "0.1.0"
-  end
-end
