+++ lib/time_iterator/core_ext/numeric.rb
