+++ lib/time_iterator/core_ext/time.rb
